docs: updating example git sort usage
Replacing the usage of the Git tags command's sort by argument `--sort=-committerdate`, which sorts by the committerdate and then reverses it.
As the preceding `-` to indicate to reverse the order does not work everywhere, e.g. Alpine Linux.

So replacing it by not reversing the order using the preceding `-` and reversing the order using tail so it is more versatile.